Quickfix: responsive UI (#10)
This pull request fixes some UI issues regarding the app responsiveness.

### This changes addresses the need by:
- Adjusting font sizes for default, `sm` and `lg` 
- Improve some margins/paddings
- Fixes a typo


Disclaimer: the breaking point for the `CloudsContainer` still has an
horizontal scroll on specific view sizes. I tried to break from
`flex-col` to `flex-row` more late but it looked strange so I think we
should keep it for now. What's your opinion on this?